Creating cash.c
Here, we calculate the maximum coin number for each kind of coin according to the money that the customer has to pay. We also want every type of coin for every customer. There are 4 kind of coins. Pennies (1 c), Nickels (5 c), Dimes (10 c) and Quarters (25 c). For exp; a customer has to pay 41 c. So, the calculation must be "25+10+5+1". Not "10+10+10+10+1" or "25+5+5+5+1". Because we want the customer to pay the price with every kind of coin. To find this calculation, we write this "cash.c" code file. Here, there is a special CS50 header file that is called "cs50.h".